Removing gRPC example

+++ src/Jason/BlazingCollatz/BlazingCollatz.ConsoleApplication/CollatzInterop.cs
@@ -1,11 +1,10 @@
 using Collatz;
-using System.Numerics;
 using System.Runtime.InteropServices.JavaScript;
 
 public partial class CollatzInterop
 {
 	[JSExport]
 	internal static int[] Generate(string start) =>
-		CollatzSequenceGenerator.Generate(BigInteger.Parse(start))
+		CollatzSequenceGenerator.Generate(int.Parse(start))
 			.Select(_ => (int)_).ToArray();
 }
